requirements.txt
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
cycler==0.10.0
ecdsa==0.17.0
et-xmlfile==1.1.0
Flask==2.0.1
Flask-AWSCognito==1.3
Flask-JWT-Extended==4.2.3
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
openpyxl==3.0.7
ordered-set==4.0.2
pdfkit==0.6.1
Pillow==8.3.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
PyLaTeX==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-jose==3.3.0
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.20
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
